Add action only remove data plugin

+++ app/Admin/Controllers/AdminPluginsController.php
@@ -74,10 +74,15 @@ class AdminPluginsController extends Controller
     {
         $key = request('key');
         $code = request('code');
+        $onlyRemoveData = request('onlyRemoveData');
         $namespace = sc_get_class_plugin_config($code, $key);
         $response = (new $namespace)->uninstall();
-        File::deleteDirectory(app_path('Plugins/'.$code.'/'.$key));
-        File::deleteDirectory(public_path('Plugins/'.$code.'/'.$key));
+
+        if(!$onlyRemoveData) {
+            File::deleteDirectory(app_path('Plugins/'.$code.'/'.$key));
+            File::deleteDirectory(public_path('Plugins/'.$code.'/'.$key));
+        }
+
         return response()->json($response);
     }
